Form augmented with auto-suggestions
User can cycle through zipcode suggestions after he/she has typed the first 3 digits of a valid zipcode. Up/down arrow keys can be used to cycle through suggestions and Enter can be used to select a zipcode and search information for it. Suggestions are clickable also.

Since any 5-digit zipcode will be automatically submitted to the server, the "submit" button was removed because it was redundant.